refactor(cpp): refactor, #113
